backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ML Model Loading
joblib==1.3.2
numpy==1.26.3
scikit-learn==1.4.0

# Utils
python-dotenv==1.0.0
